Updates to code
Added plugin path, but not plug-in capability
