+++ app/main.html
@@ -2,22 +2,24 @@
 <html lang="en" ng-app="QMCUtilities">
 
 <head>
-    <title>Snippets are cool</title>
+    <title>QMC Utilities</title>
     <meta charset="UTF-8">
     <link rel="stylesheet" type="" href="bower_components/bootstrap/dist/css/bootstrap.min.css">
     <link rel="stylesheet" type="" href="public/css/leonardo-ui.css">
     <link rel="stylesheet" type="" href="public/css/prism.css">
     <link rel="stylesheet" type="" href="public/css/menu.css">
-    <link rel="stylesheet" type="" href="app/sheetapprover/css/sheet-approver.css">
+    <link rel="stylesheet" type="" href="plugins/sheetapprover/css/sheet-approver.css">
     <script src="public/js/leonardo-ui.js"></script>
     <script src="bower_components/jquery/dist/jquery.min.js"></script>
     <script src="bower_components/bootstrap/dist/js/bootstrap.min.js"></script>
     <script src="bower_components/angular/angular.min.js"></script>
     <script src="bower_components/angular-bootstrap/ui-bootstrap.js"></script>
     <script src="bower_components/angular-route/angular-route.min.js"></script>
+    <script src="bower_components/angular-load/angular-load.js"></script>
     <script src="app/main.js"></script>
     <script src="app/menu-list.component.js"></script>
-    <script src="app/sheetapprover/sheet-approver.component.js"></script>
+    <script src="plugins/sheetapprover/sheet-approver.component.js"></script>
+    <script src="plugins/hworld/hello-world.component.js"></script>
 </head>
 
 <body>
