+++ app/main.js
@@ -1,15 +1,15 @@
 (function() {
   "use strict";
-   var module = angular.module("QMCUtilities",["ngRoute"]);
+   var module = angular.module("QMCUtilities",["ngRoute","angularLoad"]);
   
   module.config(function($routeProvider){
     $routeProvider
       .when("/:name*", {templateUrl: function(urlAttr) {
        
-        return "app/" + urlAttr.name + "/index.html";
+        return "plugins/" + urlAttr.name + "/index.html";
         
       }})
       .otherwise({ redirectTo:"/home"});
-  });
 
+  });
 }());
