+++ data/menu.json
@@ -1,32 +1,1 @@
-[
-  {
-    "id":1,
-    "title":"Governed Metrics Service",
-    "templateUrl": "/gms"
-  },
-  {
-    "id":2,
-    "title":"Sheet Approver",
-    "templateUrl": "/sheetapprover"
-  },
-  {
-    "id":3,
-    "title":"Custom Property Bulk Loader",
-    "templateUrl": "/customProp"
-  },
-  {
-    "id":4,
-    "title":"Source Control Light",
-    "templateUrl": "/sourceControl"
-  },
-  {
-    "id":5,
-    "title":"Purge Applications",
-    "templateUrl": "/purgeApps"
-  },
-  {
-    "id":6,
-    "title":"About",
-    "templateUrl": "/about"
-  }
-]
+[{"id":1,"title":"Hello World","templateUrl":"/hworld"},{"id":2,"title":"Sheet Approver","templateUrl":"/sheetapprover"}]
