+++ plugins/hworld/helloWorld.html
@@ -1,0 +1,2 @@
+<h1>{{model.response}}</h1>
+<button id="but1" class="btn-primary" ng-click="model.clickMe()">Click Me!</button>
