+++ plugins/hworld/index.html
@@ -1,0 +1,3 @@
+
+<h1>Give me something</h1>
+<hello-world></hello-world>
