+++ plugins/hworld/manifest.json
@@ -1,0 +1,4 @@
+{
+    "title":"Hello World",
+    "pluginDir": "hworld"
+}
