+++ plugins/sheetapprover/css/sheet-approver.css
