+++ plugins/sheetapprover/data/tableDef.json
@@ -1,0 +1,85 @@
+{
+    "entity": "App.Object",
+    "columns": [
+        {
+            "colNum": 1,
+            "friendlyName": "Sheet Name",
+            "display": true,
+            "colWidth":200,
+            "name": "name",
+            "columnType": "Property",
+            "definition": "name"
+        },
+        {
+            "colNum": 2,
+            "friendlyName": "Sheet Owner",
+            "display": true,
+            "colWidth":200,
+            "name": "owner",
+            "columnType": "Property",
+            "definition": "owner.name"
+        },
+        {
+            "colNum": 3,
+            "friendlyName": "Approved?",
+            "display": true,
+            "colWidth":100,
+            "name": "approved",
+            "columnType": "Property",
+            "definition": "approved"
+        },
+        {
+            "colNum": 4,
+            "friendlyName": "Published?",
+            "display": true,
+            "colWidth":100,
+            "name": "published",
+            "columnType": "Property",
+            "definition": "published"
+        },
+        {
+            "colNum": 5,
+            "friendlyName": "Last Modified",
+            "display": true,
+            "colWidth":200,
+            "name": "modifiedDate",
+            "columnType": "Property",
+            "definition": "modifiedDate"
+        },
+        {
+            "colNum": 6,
+            "friendlyName": "App",
+            "display": true,
+            "colWidth":200,
+            "name": "app.name",
+            "columnType": "Property",
+            "definition": "app.name"
+        },
+        {
+            "colNum": 7,
+            "friendlyName": "Stream",
+            "colWidth":200,
+            "display": true,
+            "name": "app.stream.name",
+            "columnType": "Property",
+            "definition": "app.stream.name"
+        },
+        {
+            "colNum": 8,
+            "friendlyName": "Sheet ID",
+            "display": true,
+            "colWidth":200,
+            "name": "id",
+            "columnType": "Property",
+            "definition": "id"
+        },
+        {
+            "colNum": 1,
+            "friendlyName": "Object Type",
+            "display": false,
+            "name": "objectType",
+            "columnType": "Property",
+            "definition": "objectType"
+        }
+    ]
+}
