+++ plugins/sheetapprover/manifest.json
@@ -1,0 +1,4 @@
+{
+    "title":"Sheet Approver",
+    "pluginDir": "sheetapprover"
+}
