+++ server.js
@@ -5,6 +5,7 @@ var fs = require('fs');
 var path = require('path');
 var config = require('./config/config');
 var winston = require('winston');
+var routeBuilder = require('./routeBuilder');
 
 //set up logging
   var logger = new (winston.Logger)({
@@ -22,7 +23,8 @@ var winston = require('winston');
   app.use('/public', express.static(config.thisServer.publicPath));
   app.use('/bower_components', express.static(config.thisServer.bowerPath));
   app.use('/data', express.static(config.thisServer.dataPath));
-  app.use('/app', express.static(config.thisServer.appPath))
+  app.use('/app', express.static(config.thisServer.appPath));
+  app.use('/plugins', express.static(config.thisServer.pluginPath));
 
   logger.info(config.thisServer.appPath);
 
@@ -30,4 +32,24 @@ var winston = require('winston');
 
   app.use('/', routes);
 
+
+  var menu = [];
+  routeBuilder.forEach(function(route, index)
+  {
+    
+    app.use('/' + route.name, require(route.path));
+    //build menu
+    var item =
+    {
+      "id": index+1,
+      "title": route.title,
+      "templateUrl": "/" + route.name
+    };
+    menu.push(item);
+  });
+
+  fs.writeFileSync(config.thisServer.dataPath + '/menu.json', JSON.stringify(menu));
+
+  
   app.listen(config.thisServer.port);
+
